Renamed utils.ts to utils/index.ts
