+++ packages/common/src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from "./validation/utils.js";
+export * from "./validation/index.js";
